Type names was upgraded

+++ InterferometricSignalsProcessing/FilterTuning/GradientTuner.cpp
@@ -8,7 +8,7 @@
 #include "FilterTuning.h"
 #include "GradientTuner.h"
 
-FilterTuning::GradientTuner::GradientTuner(float **inputSignals_, int signalSize_, int signalsCount_, int iterationsCount_,
+FilterTuning::GradientTuner::GradientTuner(double **inputSignals_, int signalSize_, int signalsCount_, int iterationsCount_,
 	ExtendedKalmanFilterIS1DState currentState_, ExtendedKalmanFilterIS1DState step_)
 	: inputSignals(inputSignals_), signalSize(signalSize_), signalsCount(signalsCount_), iterationsCount(iterationsCount_), 
 	currentState(currentState_), step(step_) {}
@@ -16,7 +16,7 @@ FilterTuning::GradientTuner::GradientTuner(float **inputSignals_, int signalSize
 
 FilterTuning::GradientTuner::~GradientTuner() {}
 
-void FilterTuning::GradientTuner::changeSignals(float **inputSignals_, int signalsCount_)
+void FilterTuning::GradientTuner::changeSignals(double **inputSignals_, int signalsCount_)
 {
 	inputSignals = inputSignals_;
 	signalsCount = signalsCount_;
@@ -29,8 +29,8 @@ void FilterTuning::GradientTuner::makeStep()
 	////Step of descent, which is influenced by variations of difference between estimation results and original signals
 	//ExtendedKalmanFilterIS1DState coef = ExtendedKalmanFilterIS1DState();
 	//ExtendedKalmanFilterIS1D filter; 
-	//dmod::signal1d recSignal(signalSize);
-	//dmod::signal1d difference(signalSize);
+	//dmod::array1d recSignal(signalSize);
+	//dmod::array1d difference(signalSize);
 
 	////Estimate varriances with descenct
 	//for (int i = 0; i < 4; i++)		//state
@@ -53,8 +53,8 @@ void FilterTuning::GradientTuner::makeStep()
 	//		{
 	//			ExtendedKalmanFilterIS1DState tmp = ExtendedKalmanFilterIS1DState();
 	//			tmp.Rw(i, j) += step.Rw(i, j);
-	//			float var_plus = FilterTuning::fitness(inputSignals, signalsCount, signalSize, currentState + tmp);
-	//			float var_minus = FilterTuning::fitness(inputSignals, signalsCount, signalSize, currentState - tmp);
+	//			double var_plus = FilterTuning::fitness(inputSignals, signalsCount, signalSize, currentState + tmp);
+	//			double var_minus = FilterTuning::fitness(inputSignals, signalsCount, signalSize, currentState - tmp);
 	//			coef.Rw(i) = gradSign(var_plus - var_minus);
 	//		}
 	//	}
@@ -67,16 +67,16 @@ void FilterTuning::GradientTuner::makeStep()
 	//		{
 	//			ExtendedKalmanFilterIS1DState tmp = ExtendedKalmanFilterIS1DState();
 	//			tmp.R(i, j) += step.R(i, j);
-	//			float var_plus = FilterTuning::fitness(inputSignals, signalsCount, signalSize, currentState + tmp);
-	//			float var_minus = FilterTuning::fitness(inputSignals, signalsCount, signalSize, currentState - tmp);
+	//			double var_plus = FilterTuning::fitness(inputSignals, signalsCount, signalSize, currentState + tmp);
+	//			double var_minus = FilterTuning::fitness(inputSignals, signalsCount, signalSize, currentState - tmp);
 	//			coef.R(i) = gradSign(var_plus - var_minus);
 	//		}
 	//	}
 	//}
 	//ExtendedKalmanFilterIS1DState tmp = ExtendedKalmanFilterIS1DState();
 	//tmp.Rn += step.Rn;
-	//float var_plus = FilterTuning::fitness(inputSignals, signalsCount, signalSize, currentState + tmp);
-	//float var_minus = FilterTuning::fitness(inputSignals, signalsCount, signalSize, currentState - tmp);
+	//double var_plus = FilterTuning::fitness(inputSignals, signalsCount, signalSize, currentState + tmp);
+	//double var_minus = FilterTuning::fitness(inputSignals, signalsCount, signalSize, currentState - tmp);
 	//coef.Rn = gradSign(var_plus - var_minus);
 
 	//currentState += coef*step;
@@ -89,7 +89,7 @@ ExtendedKalmanFilterIS1DState FilterTuning::GradientTuner::tune()
 	return currentState;
 }
 
-int FilterTuning::GradientTuner::gradSign(float s, float interval)
+int FilterTuning::GradientTuner::gradSign(double s, double interval)
 {
 	if (s > interval)
 		return 1;
