+++ InterferometricSignalsProcessing/Filters/ExtendedKalmanFilterIS1D.h
@@ -7,13 +7,13 @@ class ExtendedKalmanFilterIS1DState
 {
 public:
 	ExtendedKalmanFilterIS1DState();
-	ExtendedKalmanFilterIS1DState(Eigen::Vector4d state_, Eigen::Matrix4d R_, Eigen::Matrix4d Rw_, float Rn_);
+	ExtendedKalmanFilterIS1DState(Eigen::Vector4d state_, Eigen::Matrix4d R_, Eigen::Matrix4d Rw_, double Rn_);
 	~ExtendedKalmanFilterIS1DState();
 
 	Eigen::Vector4d state;
 	Eigen::Matrix4d R;
 	Eigen::Matrix4d Rw;
-	float Rn;
+	double Rn;
 
 	ExtendedKalmanFilterIS1DState operator+(ExtendedKalmanFilterIS1DState s);
 	ExtendedKalmanFilterIS1DState operator-(ExtendedKalmanFilterIS1DState s);
@@ -26,23 +26,23 @@ public:
 class ExtendedKalmanFilterIS1D
 {
 public:
-	ExtendedKalmanFilterIS1D(Eigen::Vector4d state_, Eigen::Matrix4d R_, Eigen::Matrix4d Rw_, float Rn_);
+	ExtendedKalmanFilterIS1D(Eigen::Vector4d state_, Eigen::Matrix4d R_, Eigen::Matrix4d Rw_, double Rn_);
 	ExtendedKalmanFilterIS1D(ExtendedKalmanFilterIS1DState full_state);
 	ExtendedKalmanFilterIS1D();
 	~ExtendedKalmanFilterIS1D();
 
 	Eigen::Vector4d getState();
 	void setState(Eigen::Vector4d st);
-	void estimate(float obs);
+	void estimate(double obs);
 	ExtendedKalmanFilterIS1DState getFullState() ;
-	float evaluateSignalValue() ;
+	double evaluateSignalValue() ;
 private:
 	Eigen::Vector4d state;
 	Eigen::Matrix4d R;
 	Eigen::Matrix4d Rw;
-	float Rn;
+	double Rn;
 	
-	float h(Eigen::Vector4d st);
+	double h(Eigen::Vector4d st);
 	Eigen::Vector4d f(Eigen::Vector4d st);
 	Eigen::Matrix4d Ft(Eigen::Vector4d st);
 	Eigen::RowVector4d Ht(Eigen::Vector4d st);
